Fix NavList structure
